requirements.txt
anthropic>=0.40.0
youtube-transcript-api>=0.6.2
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
weasyprint>=62.0
ebooklib>=0.18
markdown>=3.5
jinja2>=3.1.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pytz>=2024.1
